requirements.txt
flask
flask-sqlalchemy
flask-login
scapy
numpy
pandas
scikit-learn
tensorflow
joblib
werkzeug